Health parameters are remembered in preferences

+++ app/src/main/java/com/example/moodtrackerproject/app/EditHealthState.kt
@@ -4,6 +4,6 @@ data class EditHealthState(
     val save: Boolean = false,
     val waterNum: Int = 0,
     val stepsNum: Int = 0,
-    val sleepNum: Float = 0.0f,
+    val sleepNum: Int = 0,
     val kcalNum: Int = 0
 )
