+++ app/src/main/java/com/example/moodtrackerproject/app/HealthState.kt
@@ -4,6 +4,6 @@ data class HealthState(
     val isEdited: Boolean = false,
     val water: Int = 0,
     val steps: Int = 0,
-    val sleep: Float = 0.0f,
+    val sleep: Int = 0,
     val kcal: Int = 0,
 )
