+++ app/src/main/java/com/example/moodtrackerproject/ui/health/HealthProps.kt
@@ -6,7 +6,7 @@ data class HealthProps(
     val startEdit: () -> Unit = {},
     val water: Int = 0,
     val steps: Int = 0,
-    val sleep: Float = 0.0f,
+    val sleep: Int = 0,
     val kcal: Int = 0,
     val action: HealthScreenActions? = null
 ) {
